refactor: cleanup for the UI modifications
Needed to clean up usages for the TechDocs reader so that they work more
naturally with how the plugin is intended to behave.

Found out that the path params already intuit themselves based on the
tests, which I want to double-verify in a running environment before
pushing. Seems strange, and I'm hoping it's based on the pathRef being
used for the resource.